+++ lib/validators/specTransformer.ts
@@ -11,7 +11,18 @@ export function transform(spec: SwaggerObject): SwaggerObject {
   }
 
   for (const [definitionName, definition] of Object.entries(spec.definitions)) {
-    insertSchemaTitle(definition, definitionName)
+    insertSchemaTitle(definition, `#/definitions/${definitionName}`)
+
+    if (definition.properties) {
+      for (const [propertyName, property] of Object.entries(
+        definition.properties
+      )) {
+        insertSchemaTitle(
+          property,
+          `#/definitions/${definitionName}/properties/${propertyName}`
+        )
+      }
+    }
   }
 
   return spec